pyproject.toml
[project]
name = "forecaster"
version = "0.1.0"
description = "OO skeleton for stock forecasting"
requires-python = ">=3.10"
readme = "README.md"
authors = [{name="Your Name", email="you@example.com"}]
dependencies = [
    "pandas>=2.1",
    "numpy>=1.24",
    "scikit-learn>=1.3",
    "pyyaml>=6.0",
    "joblib>=1.3",
    "matplotlib>=3.8"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.5.0",
    "black>=24.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
select = ["E","F","I","UP","B"]
ignore = []

[tool.black]
line-length = 100

